moved generate_tree function to utils.py; created generate_tree wrappers in enrichment results
