+++ src/kinase_library/enrichment/mea.py
@@ -8,7 +8,7 @@ import os
 import pandas as pd
 import gseapy as gp
 
-from ..utils import _global_vars, exceptions
+from ..utils import _global_vars, exceptions, utils
 from ..modules import data, enrichment
 from ..objects import phosphoproteomics as pps
 from ..logger import logger
@@ -579,3 +579,45 @@ class MeaEnrichmentResults(object):
                                            title=title, xlabel=xlabel, ylabel=ylabel,
                                            plot=plot, save_fig=save_fig, return_fig=return_fig,
                                            ax=ax, **plot_kwargs)
+
+    def generate_tree(self, path, sort_by: str ='p-value', sort_direction: str = 'ascending', filter_top: int = None, **kwargs):
+        """
+        Generate a colored kinome tree from the enrichment results.
+
+        Parameters
+        ----------
+        path : str
+            Path to the kinome tree file.
+        sort_by : str, optional
+            Column name to sort the DataFrame by before generating the tree. Default is 'p-value'.
+        sort_direction : str, optional
+            Direction to sort the DataFrame, either 'ascending' or 'descending'. Default is 'ascending'.
+        filter_top : int, optional
+            If provided, only the top N rows will be included in the tree. Default is None (no filtering).
+        **kwargs : optional
+            Additional keyword arguments to be passed to the `utils.generate_tree` function.
+        """
+
+        # Check if the sort_by column exists in the enrichment results
+        if sort_by not in self.enrichment_results.columns:
+            raise ValueError(f"Column '{sort_by}' not found in enrichment results. Available columns: {self.enrichment_results.columns.tolist()}")
+
+        # Check if the sort_direction is valid
+        if sort_direction not in ['ascending', 'descending']:
+            raise ValueError("sort_direction must be either 'ascending' or 'descending'.")
+        
+    
+        # Sort the DataFrame based on the specified column and direction
+        df = self.enrichment_results.sort_values(by=sort_by, ascending=(sort_direction == 'ascending'))
+
+        if filter_top is not None:
+            df = df.head(filter_top)
+        
+        # Compute max and min NES values, ignoring NaN or non-numeric values
+        nes_values = df['NES']
+        nes_values = nes_values[pd.to_numeric(nes_values, errors='coerce').notnull()]
+        maxNes = nes_values.max()
+        minNes = nes_values.min()
+
+        # This kinome tree coloring will always be based on 'NES'
+        return utils.generate_tree(df, path, "NES", { "high": maxNes, "middle": 0.0, "low": minNes }, **kwargs)
