+++ src/kinase_library/objects/core.py
@@ -1490,147 +1490,3 @@ class Kinome(object):
             return(kin_mat_dict)
 
         return(df_kin_full_mat)
-    
-    def generate_tree(
-        self,
-        path,
-        kinase_matrix,
-        filter_top=None,
-        coloring_column="Percentile",
-        branch_color: str = "#663636",
-        node_size: int = 5,
-        low_color: str = "#999acf",
-        mid_color: str = "#c8c8c8",
-        high_color: str = "#fa6464",
-        thresholds: dict = { "high": 100, "middle": 90, "low": 60 }
-    ):
-        """
-        Generate a colored kinome tree.
-
-        Parameters
-        ----------
-        kinase_matrix : pd.DataFrame
-            DataFrame containing kinases as indices and numerical columns to color the nodes. e.g. the output of  kl.Substrate('PSVEPPLsQETFSDL').predict()
-        filter_top : int, optional
-            Number of top kinases to include in the tree. Default is None, which means all kinases will be included.
-        coloring_column : str, optional
-            Column name in the kinase matrix to use for coloring the nodes. Default is "Percentile Rank".
-        path : str
-            Path to save the tree image.
-        branch_color : str
-            Hex color for the tree branches. Default is "#663636".
-        node_size : int
-            Size of the nodes (SVG circles). Default is 5.
-        low_color : str
-            Hex color for the low end of the heatmap. Default is "#999acf".
-        mid_color : str
-            Hex color for the midpoint of the heatmap. Default is "#c8c8c8".
-        high_color : str
-            Hex color for the high end of the heatmap. Default is "#fa6464".
-        thresholds : dict
-            Dictionary containing the thresholds for low, middle, and high values. Default is { "high": 100, "middle": 90, "low": 60 } for "Percentile".
-        """
-
-        def hex_to_rgb(hex_color):
-            """Convert hex string like '#FF0000' to (255, 0, 0)"""
-            hex_color = hex_color.lstrip('#')
-            return tuple(int(hex_color[i:i+2], 16) for i in (0, 2, 4))
-
-
-        def rgb_to_hex(rgb):
-            """Convert (255, 0, 0) to '#FF0000'"""
-            return '#{:02X}{:02X}{:02X}'.format(*rgb)
-
-
-        def calculate_color(value, high_color_hex, low_color_hex, high_value, low_value):
-            high_color = hex_to_rgb(high_color_hex)
-            low_color = hex_to_rgb(low_color_hex)
-
-            new_value = max(min(value, high_value), low_value)
-            new_value -= low_value
-
-            percentage = new_value / (high_value - low_value) if high_value != low_value else 0
-
-            red = high_color[0] * percentage + low_color[0] * (1.0 - percentage)
-            green = high_color[1] * percentage + low_color[1] * (1.0 - percentage)
-            blue = high_color[2] * percentage + low_color[2] * (1.0 - percentage)
-
-            return rgb_to_hex((round(red), round(green), round(blue)))
-
-
-        def calculate_heatmap_color_midpoint(value, high_color, mid_color, low_color, high_value, mid_point_value, low_value):
-            if value < mid_point_value:
-                return calculate_color(value, mid_color, low_color, mid_point_value, low_value)
-            return calculate_color(value, high_color, mid_color, high_value, mid_point_value)
-
-
-        def calculate_heatmap_color(value, high_value, mid_point_value, low_value, high_color, mid_color, low_color):
-            return calculate_heatmap_color_midpoint(
-                value,
-                high_color,
-                mid_color,
-                low_color,
-                high_value,
-                mid_point_value,
-                low_value
-            )
-
-        # Check if path is valid
-        if not isinstance(path, str) or not path.endswith('.svg'):
-            raise ValueError("Path must be a valid string ending with '.svg'.")
-        
-        if kinase_matrix.get(coloring_column, None) is None:
-            raise ValueError(f"Column '{coloring_column}' not found in the kinase matrix. Please provide a valid column name.")
-
-        # Check to make sure that the top X kinases is less than number of kinases
-        if filter_top is not None and filter_top <= len(kinase_matrix):
-            top_kinases = kinase_matrix.nlargest(filter_top, coloring_column).index
-        elif filter_top is None:
-            top_kinases = kinase_matrix.index
-        else:
-            raise ValueError("filter_top must be less than or equal to the number of kinases.")
-        
-        # Create map for quick access to kinase names
-        kinase_uniprot_mapping = {
-            row['UNIPROT_ID']: row['MATRIX_NAME']
-            for _, row in data.get_kinome_info().iterrows()
-        }
-
-        # Load the base SVG
-        import xml.etree.ElementTree as ET
-        current_dir = os.path.dirname(__file__)
-        svg_path = os.path.abspath(os.path.join(current_dir, "../databases/kinase_data/base_tree.svg"))
-        tree = ET.parse(svg_path)
-        root = tree.getroot()
-
-        # Set the branch color if not default
-        if branch_color != "#663636":
-            for line in root.findall('.//svg:path', namespaces={'svg': 'http://www.w3.org/2000/svg'}):
-                line.set('fill', branch_color)
-
-        ns = {'svg': 'http://www.w3.org/2000/svg'}
-
-        # For the top X kinases in kinase matrix, set the opacity to 1
-        for circle in root.findall('.//svg:circle', namespaces=ns):
-            uniprot_id = circle.get('class').split("_")[-1]
-            if kinase_uniprot_mapping.get(uniprot_id, None) in top_kinases:
-                circle.set('opacity', '1')
-
-        for circle in root.findall('.//svg:circle', namespaces=ns):
-            uniprot_id = circle.get('class').split("_")[-1]
-            if kinase_uniprot_mapping.get(uniprot_id, None) in top_kinases:
-                val = kinase_matrix.at[kinase_uniprot_mapping.get(uniprot_id, None), coloring_column]
-                if val is not None:
-                    color = calculate_heatmap_color(val, thresholds["high"], thresholds["middle"], thresholds["low"],
-                                                    high_color, mid_color, low_color)
-                    circle.set('fill', color)
-                    circle.set('stroke', "gray")
-                    circle.set('stroke-width', "0.5px")
-                    circle.set('opacity', '1')
-                    circle.set('r', str(node_size))
-
-        try:
-            os.makedirs(os.path.dirname(path), exist_ok=True)
-            tree.write(path)
-        except Exception as e:
-            raise Exception(f"Error saving SVG file: {e}")
